Magic Missile Starting
delayedCalls within gameobjects (Player, MagicMissile, MagicMissileBlast) never go off, and that needs to be fixed.

collisionChecking within gameobjects (MagicMissile, MagicMissileBlast) is behaving strangely, maybe put it in the scene. Actually wait what if there was a Projectiles class that these things could inherit from????

detonate(), the method that explodes the MagicMissile into MagicMissile blast, can't remember MagicMissile. It's super weird.

Try and fanagle the collisions for a little, and maybe carry a reference to the blast? idk, I'm gonna sleep on it.